requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
psycopg2-binary==2.9.5
python-decouple==3.7
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
stripe==5.0.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
gunicorn==20.1.0